+++ app/lib/errors.ts
@@ -1,0 +1,3 @@
+export class ConflictError extends Error {}
+export class NotFoundError extends Error {}
+export class InternalError extends Error {}