+++ app/lib/queries.ts
@@ -2,6 +2,8 @@ import {prisma} from '@/app/lib/prisma';
 import {TweetWithAuthor} from '@/app/lib/definitions';
 import {generateHandle, getAvatarForHandle} from '@/app/lib/utils';
 import bcrypt from 'bcryptjs';
+import {PrismaClientKnownRequestError} from '@/generated/prisma/internal/prismaNamespace';
+import {ConflictError, NotFoundError} from '@/app/lib/errors';
 
 export async function getTweetsWithAuthor(): Promise<TweetWithAuthor[]> {
     const tweets = await prisma.tweet.findMany({
@@ -76,4 +78,51 @@ export async function createUser(data: {
             handle: true,
         },
     });
+}
+
+export async function getUser(handle: string) {
+    const user = prisma.user.findUnique({where: {handle}});
+    if (!user) {
+        throw new Error('No  user with the handle ' + handle + ' found');
+    }
+    return user;
+}
+
+export async function searchUsers(query: string) {
+    return prisma.user.findMany({
+        where: {
+            OR: [
+                {handle: {contains: query, mode: "insensitive"}},
+                {name: {contains: query, mode: "insensitive"}},
+                {username: {contains: query, mode: "insensitive"}},
+            ]
+        },
+        select: {
+            id: true,
+            name: true,
+            username: true,
+            handle: true,
+            avatar: true,
+        },
+        take: 10
+    })
+}
+
+export async function updateHandle(handle: string, newHandle: string) {
+    try {
+        return await prisma.user.update({
+            where: {handle},
+            data: {handle: newHandle},
+        })
+    } catch (error) {
+        if (error instanceof PrismaClientKnownRequestError) {
+            if (error.code === 'P2002') {
+                throw new ConflictError('Handle already exists');
+            }
+            if (error.code === 'P2025') {
+                throw new NotFoundError('User not found');
+            }
+        }
+        throw error
+    }
 }