+++ prisma/migrations/20260607225825_fix_account_unique_and_refresh_token/migration.sql
@@ -1,0 +1,13 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `refres_token` on the `Account` table. All the data in the column will be lost.
+  - A unique constraint covering the columns `[provider,providerAccountId]` on the table `Account` will be added. If there are existing duplicate values, this will fail.
+
+*/
+-- AlterTable
+ALTER TABLE "Account" DROP COLUMN "refres_token",
+ADD COLUMN     "refresh_token" TEXT;
+
+-- CreateIndex
+CREATE UNIQUE INDEX "Account_provider_providerAccountId_key" ON "Account"("provider", "providerAccountId");