+++ proxy.ts
@@ -1,1 +1,14 @@
-export {auth as proxy} from "@/auth";
+import {auth} from '@/auth';
+import {NextResponse} from 'next/server';
+
+export default auth((req => {
+    if (!req.auth) {
+        return NextResponse.redirect(new URL("/auth", req.url));
+    }
+}))
+
+export const config = {
+    matcher: [
+        "/(protected)/:path*"
+    ]
+}